requirements.txt
python-dotenv
requests
pillow
matplotlib
ipykernel
pandas